requirements.txt
# Telegram Bot
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.10.11

# API
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1

# Data processing
pandas==2.2.3

# Reports generation
reportlab==4.2.5
weasyprint==62.3
openpyxl==3.1.5

# Additional
python-multipart==0.0.12
pydantic==2.9.2
